Implementacion de Gates

+++ app/Http/Controllers/BookController.php
@@ -7,6 +7,7 @@ use App\Models\User;
 use App\Models\Book;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Gate;
 
 class BookController extends Controller
 {
@@ -40,6 +41,8 @@ class BookController extends Controller
      */
     public function create()
     {
+       Gate::authorize('admin-book');
+
         $publishers = Publisher::get();
         return view('book.bookForm', compact('publishers'));
         //return view('book.bookForm');
@@ -53,6 +56,8 @@ class BookController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('admin-book');
+
         $request->validate($this->rules);
         Book::create($request->all());
         return redirect()->route('book.index');
@@ -78,6 +83,8 @@ class BookController extends Controller
      */
     public function edit(Book $book)
     {
+        Gate::authorize('admin-book');
+
         $publishers = Publisher::get();
         return view('book.bookForm', compact('book','publishers'));
     }
@@ -91,6 +98,8 @@ class BookController extends Controller
      */
     public function update(Request $request, Book $book)
     {
+        Gate::authorize('admin-book');
+
         $request->validate($this->rules);
         Publisher::where('id',$book->id)->update($request->except('_token','_method'));
         return redirect()->route('book.show', $book);
@@ -104,6 +113,8 @@ class BookController extends Controller
      */
     public function destroy(Book $book)
     {
+        Gate::authorize('admin-book');
+
         $book->delete();
         return redirect()->route('book.index');
     }
@@ -116,6 +127,8 @@ class BookController extends Controller
      */
     public function addUser(Request $request, Book $book)
     {
+        Gate::authorize('admin-book');
+
         $book->users()->sync($request->user_id);
         return redirect()->route('book.show', $book);
     }
